Extract test_agent_chains.py
